Rename plugin and bot directories
