+++ DemiCatPlugin/ChatWindow.cs
@@ -8,7 +8,7 @@ using System.Text.RegularExpressions;
 using System.Numerics;
 using Dalamud.Bindings.ImGui;
 
-namespace DalamudPlugin;
+namespace DemiCatPlugin;
 
 public class ChatWindow : IDisposable
 {
