+++ DemiCatPlugin/EmbedDto.cs
