+++ DemiCatPlugin/manifest.json
@@ -1,7 +1,7 @@
 {
   "Author": "DemiCat",
   "Name": "DemiCat",
-  "InternalName": "DemiCat",
+  "InternalName": "DemiCatPlugin",
   "AssemblyVersion": "1.1.0.0",
   "Description": "DemiCat Dalamud plugin.",
   "ApplicableVersion": "any",
