+++ discord-demibot/.env.example
