+++ discord-demibot/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "bot",
+  "name": "discord-demibot",
   "version": "1.0.0",
   "description": "",
   "main": "src/index.js",
