+++ discord-demibot/src/http/routes/admin/setup.js
