+++ discord-demibot/src/logger.js
